Address mentor's feedback:
kotp
5h ago

The style that is better is likely to not need to iterate the collection twice.

Even better is likely to write the collection as it should be, instead of needing to transform it.

Fortunately, you have made this easy, since we could use the code to write what we "really mean".

If we call IRB` and load this file, then we can gain access to a representation of that constant.

Or we could programmatically append it to this file, even.

Whichever clever way you might like to do this, but whatever way, may not be apparent in the result.

The argument list of input is valid, but is also valid for any method that takes a single argument. It is valid, but not informative. It does not answer the question of "What kind of input." This will likely drive the reader of the code to make assumptions (which may be right, but why not give a bit better help) by other areas of the code, or heaven forbid, from the documentation that was generated (only to get "input" as part of that documentation, and now they have to go to the non-generated documentation, the README).

Instead, if that one word were changed to color, then the question disappears, while still communicating "we need input".

If we want to be clever, then we might build a string directly rather than using map and we would also not need to join the Array we are creating, as we would already have a String.

More cleverly, though, we might take the index of the input in reverse and power the index by ten, and the multiply it by the decoded value.

This feedback is based on the desire to be clever, as indicated in something you wrote, outside of what is presented here.

I hope it helps.

I am looking forward to the next iteration!