Refac using 'initialize' instead of 'new'. fix typo
typo with constant string '<name>'

Also, note that there is no default in the instance level call, which is very nice. If a name is required for the "power user" then that is great.

The class level method is there for convenience only, and it should stay inflexible, as it is there for convenience not power. We will see this decision in use in our "expansion", below.