REFAC: attr_rdr, unfreeze CONST and add languages
Victor Goff's(kotp) comments:
Now we can set up a second argument for our argument list, but where do we do this? As mentioned, our convenience method at the class level should remain convenient and not flexible.

So this means we are likely going to to add to our argument list in initialize where we can give English as the default but allow for the user to provide another language which we maintain and support.

However, we should not also support this while making it harder for the user of our library to support themselves for languages that we do not maintain.

This last thing really brings me to the words that are used to describe answers to "Can I use this with X?". Sometimes you will hear "It is not supported." This answer should mean "Of course you can do it, but we do not support it, if you run into a problem, do not come to us with a fix."

The "it is not supported" is not an answer to "can we do it?".

So let's provide the mechanisms for what we do support, even if we do not maintain the languages, and allow this to be used in ways easily, without a whole lot of work.