privatize constant, remove 'then'
The then is a Kernel method, and helpful to know. But without a block it created an Enumerator object, with a block it does what an Enumerable would do. We would usually use it when the object we are calling it on is not an Enumerable object.

But really we do not need to use any iterating, if we deconstruct what we have for what we want in the body of the block.

    first_color, second_color = color_pair
    @value = FORMAT % { tens: BAND[first_color], ones: BAND[second_color] }

I do not think it makes sense to freeze the FORMAT string as a singleton object, it is the only one anyway. And if a user wanted to change how it looks, let them? The class is open and they could do so anyway, so unless we are trying to make it harder for them, why make it harder for them to do so?

If it is not meant to be used by the user of the library, then make it private (rather than freezing it, different communication).