relocate test, make class method less flexible. rm frzn_strng
Feedback:
I would not freeze the string literals. There are none here that should be frozen, and really not an overabundance of strings being created, since they are referenced in a constant.

If the user of the library needs to make a change for their purposes, those constants are similar to the constant TwoFer, which is not frozen, but open for modifications, changes, state differences, etc.

The 'you' string is not really subject to modification, malicious or other wise, since it is a local variable assignment, and only if it is not replaced and also only through the convenience method.

The class level method, we should not have the flexibility.

And the "given" class level method should not be changed, yet was changed to:

  def self.two_fer(name = 'you', language = :English)
    new(name, language).to_s
  end

It should be left as a "given" as a proven solution, that we will complete with the rest of our code.

The class level method is here for convenience, without flexibility, and so should be as:

def self.two_fer(name = 'you')
  new(name).to_s
def

If changes are wanted, we should discuss them first. Otherwise we are changing more than we should in the exploration, changes that are happening at two different levels, and it will not be helpful to either of us. I will not be able to easily track and see the reasoning that something changed, and we will not have discussed the too many moving parts to know which way we should go.

Instead of submitting changes to the tests, we can show example runs by sharing a "Library Guard" at the bottom of the submission.

It looks like this:

if $PROGRAM_NAME == __FILE__
  puts TwoFer.two_fer('José')
  puts TwoFer.new('Joseph')
  puts TwoFer.new('José', language: :Español)
end

The output for the above should end up being this:

One for José, one for me.
One for Joseph, one for me.
Uno para José, uno para mi.