REFAC: setup data in inititializer
Commets before this commit:

If we fully construct the value in our initialize, then we probably do not need to have the attributes for first_color and second_color.

We probably have not "normalized our internal state" far enough, if we are having to determine the value in this way all the way to to_s.

Here is the diff with my "newline inside class definition" and how I see the initialize normalizing to the point where we do not have to carry too much internal state.
Click here to see the diff

```ruby
 require_relative 'band'

 class ResistorColorDuo
+
   DISPLAY = '%<tens>s%<ones>s'

   def self.value(colors)

   private

-  attr_reader(
-    :first_color,
-    :second_color,
-    :value
-  )
+  attr_reader :value

   def initialize(color_pair)
-    @first_color, @second_color = color_pair
-    @value = to_s
+    first_color, second_color = color_pair
+    @value = DISPLAY % { tens: BAND[first_color], ones: BAND[second_color] }
   end

   def to_s
-    DISPLAY % { tens: BAND[first_color], ones: BAND[second_color] }
+    value
   end

   public
@@ -29,4 +26,5 @@ class ResistorColorDuo
   def to_i
     value.to_i
   end
+
 end

Now it is time to talk about the Elephant in the room.

The band.rb file.

This constant is in the main scope, it is polluting the namespace, and when I wrote my musical band software for a bunch of electricians, it keeps stomping on my class name. (Mostly because I named it BAND.)

Still, we can be a bit more polite, by namespacing our BAND constant.

Would you want to throw that into a module, maybe? Perhaps Resistor is a good name for a module that would have things pertaining to resistors?