Remove redundant constants for indiviadual translations
based on comment:
OK, but nothing was changed in this regard, we still have:

  ENGLISH = 'One for %<name>s, one for me.'
  ESPAÑOL = 'Uno para %<name>s, uno para mi.'
  ESPARANTO = 'Unu por %<name>s, unu por mi.'

  LANGUAGE = {
    English: ENGLISH,
    Español: ESPAÑOL,
    Esparanto: ESPARANTO
  }

Which is 4 different things, while the LANGUAGE should be the source of truth for the translations, rather than the individual constants. Otherwise we are mapping a name (symbol, which is literally important because of its identity) and a reference to which constant is supposedly storing the important information.

Let us instead tie this together concretely by moving the strings for the those constants down to the keys that we want to use. We can then get rid of the three language specific constants.

Now the keys will be directly tied to what we want, without a potential disconnect.