Generade rdoc html docs using 'rdoc ./'
This uses a gem called rubydoctest and has the benefit of encouraging documentation, while giving IRB session syntax, but also being executable code, so that we know when our code examples are not truthful, since they will break if they are not truthful to the way to use the library.
If you are developing locally, and hopefully you are, run rdoc . and then checkout the documentation that is generated in the ./doc directory.